requirements.txt
tqdm
pydantic
pypdf
rich
entence-transformers 
faiss-cpu 
numpy